Support 2 channels with softmax for binary segmentation(2 classes)

+++ datasets/spacenet6optical.py
@@ -46,11 +46,11 @@ class SpaceNet6Optical(Dataset):
 
         augmentations = self.transform(image=image, mask=mask)
         image = augmentations['image'].float()
-        mask = augmentations['mask'].long()
+        mask = augmentations['mask'].float()
 
         data = {
-            'image': image.float(),
-            'label': mask.float(),
+            'image': image,
+            'label': mask,
             'image_path': image_path
         }
